feat: refactor getCentroid
