+++ src/__tests__/index.test.ts
@@ -4,7 +4,7 @@ import { toBeDeepCloseTo } from 'jest-matcher-deep-close-to';
 import { ransac } from '..';
 import { applyRigidTransfom } from '../rigidTransform/applyRigidTransform';
 import { createRigidTransformModel } from '../rigidTransform/createRigidTransformModel';
-import { drawResult } from '../rigidTransform/drawResult';
+import { drawResult } from '../toMigrate/forImageJs/drawResult';
 import { getEuclidianDistance } from '../rigidTransform/getEuclidianDistance';
 import { getRigidTransformArray } from '../rigidTransform/getRigidTransform';
 import { line, linearRegression } from '../utils/linearRegression';
