+++ src/rigidTransform/README.md
@@ -4,6 +4,8 @@ Based on the tutorial: https://nghiaho.com/?page_id=671
 
 ## Points format
 
+This is a format specific to `image-js` and should not be in the code to get the affine transform. We'd rather directly use matrices.
+
 ```js
 /**
  * Coordinates of a point in an image with the top-left corner being the reference point.
