+++ src/toMigrate/forImageJs/drawResult.ts
