+++ src/toMigrate/forImageJs/getMatrixFromPoints.ts
