+++ src/toMigrate/forImageJs/getPointsFromMatrix.ts
